requirements.txt
python-multipart==0.0.18
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlmodel==0.0.22
python-dotenv==1.0.1
httpx==0.27.2
Pillow==11.0.0
openai==1.54.3
